Materials and Orthogonal
Made most materials triplanar so that they can be the same for any mesh size. This allowed separating the meshes from the materials, so there's now a resource file for each plane mesh shape.
Also added a level that involves a new mechanic which will likely be used in some late-game levels.