Bulb on/off and Door Interact
Added a script to the ceiling bulb light, enabling the light to be turned on and off, as well as allowing variations in hight and size.
Made the door interact outline work on connected doors, so that when looking through a door, the door on the other side will be outlined.